DDCNL-9021 | separated transfer view tests into separate files
